requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.1/constraints-3.11.txt"
apache-airflow-providers-cncf-kubernetes==8.4.1
Jinja2==3.1.4
kubernetes==30.1.0
PyYAML==6.0.2
